requirements.txt
torch>=1.9.0
numpy>=1.21.0
scipy>=1.7.0
h5py>=3.1.0
PyYAML>=5.4.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0
pandas>=1.3.0
